pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.60.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.12.2",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
]
